requirements.txt
fastapi
uvicorn
onnxruntime
pillow
numpy
python-multipart
mediapipe